HTR - VIt Reference/requirements.txt
appdirs==1.4.4
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
editdistance==0.8.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gin==0.1.6
gin-config==0.5.0
gitdb==4.0.11
gitpython==3.1.43
google-auth==2.34.0
google-auth-oauthlib==1.0.0
grpcio==1.66.1
huggingface-hub==0.24.6
idna==3.7
imageio==2.34.1
importlib-metadata==8.4.0
importlib-resources==6.4.4
joblib==1.4.0
kiwisolver==1.4.7
lazy-loader==0.4
markdown==3.7
markupsafe==2.1.5
matplotlib==3.7.5
networkx==3.1
nltk==3.8.1
numpy==1.24.4
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
pip==24.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1-modules==0.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pywavelets==1.4.1
pyyaml==6.0.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.4
scikit-image==0.21.0
scipy==1.10.1
sentry-sdk==2.0.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tifffile==2023.7.10
timm==1.0.9
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchsummary==1.5.1
torchvision==0.14.0+cu116
tqdm==4.65.2
typing-extensions==4.11.0
urllib3==2.2.1
wandb==0.16.6
werkzeug==3.0.4
zipp==3.20.1
